Added initial project
The app has Rollify.Core, which holds platform-independent code,
like database logic (which hooks into a native implementation) and
dice rolling logic. The app takes a mathetematical expression
(which can include dice definitions) and evaluates it, rolling any
dice necessary. It keeps a database of formulas with an expression
and a name. Currently the name is the same as the expression, but
a menu will be added to create the formula in the future.