Separated core app logic from ui
Core now maintains the databases and provides handles for the essential
functions to allow user interface code to call them. This allows
platform-specific code to focus only on providing a user interface and
platform interface, with shared code that defines what to do when events
are triggered. This will make cross-platform support much easier in the
future, and willl also allow easier multithreading. Because the UI is
separated from logic, it's easy to separate the threads as well.