Added some more roller logic
The last commit wasn't quite complete; now it works. It's faster because
it doesn't parse the expression every time. It just reads it once and
builds a code structure modeling what the expression is supposed to do.
The are probably lots more optimizations to make, but this was a big
one.